added minimum tolerance for newton iterations.
About the C parameter "The reason for introducing it is as follows.  For a given outer iteration (i.e. fixed t) the precision of the result will be ~ m/t  (in terms of how close it is to the solution of the original inequality constrained problem).  It is possible to run Newton solve to a fixed high precision (say 1e-12) to get this approximate solution, but this si wasteful.  On the other hand, running it at the same precision as the current error due to finite t is somewhat risky, especially for low values of t -- this may lead to a substantial deviation from the true solution.    So a possible approach is to use a faction c of the current outer iteration precision.  This however is excessive for high values of t, especially hitting the limits of double accuracy.  It is reasonable to limit how high we set Newton accuracy. We can try either  max(  c*m/t,  e_b)  or max( c*m/t, e_max)  with e_max being the max accuracy we hope to achieve with doubles in the Neuton solve (say 1e-12 or 1e-11)."