Termination (#22)
* Added new termination criterias

* changed newton lineseach stopping criteria

Copy e_b from barrier solver to newton solver.

* added minimum tolerance for newton iterations.

About the C parameter "The reason for introducing it is as follows.  For a given outer iteration (i.e. fixed t) the precision of the result will be ~ m/t  (in terms of how close it is to the solution of the original inequality constrained problem).  It is possible to run Newton solve to a fixed high precision (say 1e-12) to get this approximate solution, but this si wasteful.  On the other hand, running it at the same precision as the current error due to finite t is somewhat risky, especially for low values of t -- this may lead to a substantial deviation from the true solution.    So a possible approach is to use a faction c of the current outer iteration precision.  This however is excessive for high values of t, especially hitting the limits of double accuracy.  It is reasonable to limit how high we set Newton accuracy. We can try either  max(  c*m/t,  e_b)  or max( c*m/t, e_max)  with e_max being the max accuracy we hope to achieve with doubles in the Neuton solve (say 1e-12 or 1e-11)."

* added fixture-script to test different eb values.

* add count of hessian and function evaluations

* adding hpc scripts

* Make last outer iteration use exactly t = m/e_b

- Make last outer iteration use exactly t = m/e_b
- Added enegy to logs (only when DEBUG_LINESEARCH is on)

* Updated generation scripts with new barrier params

* add ENABLE_DERIVATIVE_CHECK to eval_g_and_gdiff

* fixed linesearch condition

* Fixed check that compares the two autodiff gradients

Fixed check that compares the two autodiff gradients
- it was using diff.norm() to compare but that doesn't work for large gradients.
- changed to use the proper function we have for comparing jacobians, but with a smaller test_eps (the two matrices should really be the same)

* Added min_distance to log an json simulation file

- remove the ifdef from debug_min_distances function
- override simulation_step and take_step to compute min_distance (maybe an overkill for simulation step)

* Added CoR = 1e-6 and Δt = 1e-1, 1e-2, 1e-3

* Added spacing to JSON by default (excluded it previously because I
    didn't like the formatting of Python's JSON)

* Fixed bug in python scripts and old unmerged

* added min_distance to barrier and changed functional to E + 1/t B

- updated barrier evaluation to use \phi(distance - min_distance)
- update simulation_step function to check if current_min_distance < min_distance. If true, then we run optimization

- updated barrier functional to use E + 1/t B